+++ backend/package.json
@@ -12,8 +12,8 @@
     },
     "license": "ISC",
     "dependencies": {
+        "@json2csv/plainjs": "^7.0.3",
         "dayjs": "^1.11.10",
-        "json2csv": "^6.0.0-alpha.2",
         "jsonfile": "^6.0.1"
     },
     "devDependencies": {
