+++ backend/packages/Upgrade/src/api/services/StratificationService.ts
@@ -3,6 +3,7 @@ import { OrmRepository } from 'typeorm-typedi-extensions';
 import { UpgradeLogger } from '../../lib/logger/UpgradeLogger';
 import { SERVER_ERROR } from 'upgrade_types';
 import { In, getConnection } from 'typeorm';
+import { Parser } from '@json2csv/plainjs';
 import { FactorStrata, StratificationInputValidator } from '../controllers/validators/StratificationValidator';
 import { ExperimentUser } from '../models/ExperimentUser';
 import { StratificationFactor } from '../models/StratificationFactor';
@@ -83,12 +84,16 @@ export class StratificationService {
   public async getCSVDataByFactor(factor: string, logger: UpgradeLogger): Promise<any> {
     logger.info({ message: `Download CSV stratification by factor. Factor: ${factor}` });
 
-    return await this.stratificationFactorRepository
+    const data = await this.stratificationFactorRepository
       .createQueryBuilder('sf')
       .select(['usf.user AS uuid', `usf.stratificationFactorValue AS "${factor}"`])
       .leftJoin('sf.userStratificationFactor', 'usf')
       .where('sf.stratificationFactorName = :factor', { factor })
       .getRawMany();
+
+    // Convert JSON data to CSV
+    const parser = new Parser();
+    return parser.parse(data);
   }
 
   public async deleteStratification(factor: string, logger: UpgradeLogger): Promise<StratificationFactor> {
@@ -97,10 +102,40 @@ export class StratificationService {
     return await this.stratificationFactorRepository.deleteStratificationFactorByName(factor, logger);
   }
 
-  public async insertStratification(
-    userStratificationData: StratificationInputValidator[],
-    logger: UpgradeLogger
-  ): Promise<UserStratificationFactor[]> {
+  public async insertStratification(csvData: string, logger: UpgradeLogger): Promise<UserStratificationFactor[]> {
+    // const csvData = request.body[0].file;
+    const rows = csvData.replace(/"/g, '').split('\n');
+    const columnNames = rows[0].split(',');
+
+    let userStratificationData: StratificationInputValidator[] = [];
+
+    // Iterate through the rows (skip the header row)
+    for (let i = 1; i < rows.length; i++) {
+      const row = rows[i];
+      const rowValues = row.split(',');
+
+      // Extract the user ID
+      const userId = rowValues[0];
+      if (!userId) {
+        continue;
+      }
+
+      // Iterate through other columns (factors)
+      for (let j = 1; j < columnNames.length; j++) {
+        const factorName = columnNames[j];
+        const factorValue = rowValues[j].trim();
+
+        // Create an object and add it to the array
+        const userFactorValue: StratificationInputValidator = {
+          userId: userId,
+          factor: factorName.trim(),
+          value: factorValue === '' ? null : factorValue,
+        };
+
+        userStratificationData.push(userFactorValue);
+      }
+    }
+
     logger.info({ message: `Insert stratification => ${JSON.stringify(userStratificationData, undefined, 2)}` });
 
     const createdStratificationData = await getConnection().transaction(async (transactionalEntityManager) => {
